Major Refactoring
Moved the looping logic inside the PasteBin Scraper, and accidentally created a better loop for grabbing the recent pastes. Instead of getting recent pastes from recursive pastes we've serializing, just ping the index page when we run out of items. Also fixed the file saving and removed the need for threading.